replace statistic service logic in presenter
